Pass vhost-user-blk device to container
As kubernetes doesn't take vhost-user-blk device into consider, when
process block volume it expects the device CSI driver provided is a
real block device. But after setup vhost-user block device, there is
only a socket and a block device file (no real block device is attached
to it). If pass the vhost-user-blk device directly to kubernetes it
will fail at device mapping stage.

To avoid this problem:
  In CSI driver:
    1. create a small image file <vhost_user_store_path>/<devname>.img
    2. losetup -f <vhost_user_store_path>/<devname>.img
    3. pass the loopback device attached to the image file to kubernetes.

  In kata, check the devices when create a new container, if the device
is a loopback device try to retieve the vhost-user-blk information and
replace the loopback device with vhost-user-blk device.