docs: add a how-to to help users to run DAX
Kata supports DAX but not out of the box.
Currently, users need to build an experimental QEMU to use
DAX with virtiofs, but the configuration file makes users confused.

This commit adds comments in the configuration file about the
limitation of DAX support and adds a how-to to help user to
build an experimental QEMU.

Fixes: #4636

Signed-off-by: liubin <liubin0329@gmail.com>
Signed-off-by: Bin Liu <bin@hyper.sh>