container-monitor: Introduction of tool for gathering conatiner stats
container-monitor is created to facilitate exposing container specific
stats which would typically be exposed by CADvisor in a kubernetes
environment. These are already available via task service API today, but
are not exposed through CRI interface.

Because of this, we'll introduce a tool that will present a prometheus
endpoint, and a Collector for gathering stats. Each time Collect is
invoked, we will get a list of containers from containerd, and then
connect to each container's shim in order to get stats. We then present
these metrics with the labels namespace, container-name and pod-name.

These labels are what a kubernetes user would see (ie, what's presented
"north" of CRI).

Usage:

$ ./container-monitor --help
NAME:
   container-monitor - A new cli application

USAGE:
   container-monitor [global options] command [command options]
[arguments...]

COMMANDS:
   help, h  Shows a list of commands or help for one command

GLOBAL OPTIONS:
   --address value, -a value      path to the containerd socket (default: "/run/containerd/containerd.sock")
   --namespace value, --ns value  the namespace to get container stats from (default: "default")
   --server-port value, -p value  The address the server listens on for HTTP requests. (default: "8090")
   --log-level value              Logging level (trace/debug/info/warn/error/fatal/panic). (default: "warn")
   --help, -h                     show help

Signed-off-by: Eric Ernst <eric_ernst@apple.com>