agent: change `AGENT_CONFIG`'s lazy type to just `AgentConfig`
Since it is never modified, it doesn't really need a lock of any kind.
Removing the `RwLock` wrapper allows us to remove all `.read().await`
calls when accessing it.

Additionally, `AGENT_CONFIG` already has a static lifetime, so there is
no need to wrap it in a ref-counted heap allocation.

Fixes #5409

Signed-off-by: Wedson Almeida Filho <walmeida@microsoft.com>